# docs/swagger.yaml
definitions:
...
  types.Column:
    properties:
      exp:
        description: expressions, which default to = when the value is null, have
          =, !=, >, >=, <, <=, like
        type: string
      logic:
        description: logical type, defaults to and when value is null, only &(and),
          ||(or)
        type: string
      name:
        description: column name
        type: string
      value:
        description: column value
    type: object
  types.Conditions:
    properties:
      columns:
        description: columns info
        items:
          $ref: '#/definitions/types.Column'
        type: array
    type: object
...
  types.CreateRoleRequest:
    properties:
      admin:
        type: string
      createBy:
        type: integer
      dataScope:
        type: string
      flag:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        type: integer
    type: object
  types.CreateRoleRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          id:
            description: id
            type: integer
        type: object
      msg:
        description: return information description
        type: string
    type: object
...
  types.DeleteRoleByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
  types.DeleteRolesByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.DeleteRolesByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
...
  types.GetRoleByConditionRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          role:
            $ref: '#/definitions/types.RoleObjDetail'
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.GetRoleByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          role:
            $ref: '#/definitions/types.RoleObjDetail'
        type: object
      msg:
        description: return information description
        type: string
    type: object
...
  types.ListRolesByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.ListRolesByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          roles:
            items:
              $ref: '#/definitions/types.RoleObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.ListRolesRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          roles:
            items:
              $ref: '#/definitions/types.RoleObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
...
  types.Params:
    properties:
      columns:
        description: query conditions
        items:
          $ref: '#/definitions/types.Column'
        type: array
      page:
        description: page number, starting from page 0
        type: integer
      size:
        description: lines per page
        type: integer
      sort:
        description: sorted fields, multi-column sorting separated by commas
        type: string
    type: object
  types.RoleObjDetail:
    properties:
      admin:
        type: string
      createBy:
        type: integer
      createdAt:
        type: string
      dataScope:
        type: string
      flag:
        type: string
      remark:
        type: string
      roleId:
        type: string
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        type: integer
      updatedAt:
        type: string
    type: object
...
  types.UpdateRoleByIDRequest:
    properties:
      admin:
        type: string
      createBy:
        type: integer
      dataScope:
        type: string
      flag:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        type: integer
    type: object
  types.UpdateRoleByIDRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: http server api docs
  title: admin api docs
  version: "2.0"
paths:
...
  /api/v1/role:
    post:
      consumes:
      - application/json
      description: submit information to create role
      parameters:
      - description: role information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CreateRoleRespond'
      security:
      - BearerAuth: []
      summary: create role
      tags:
      - role
  /api/v1/role/{id}:
    delete:
      consumes:
      - application/json
      description: delete role by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteRoleByIDRespond'
      security:
      - BearerAuth: []
      summary: delete role
      tags:
      - role
    get:
      consumes:
      - application/json
      description: get role detail by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetRoleByIDRespond'
      security:
      - BearerAuth: []
      summary: get role detail
      tags:
      - role
    put:
      consumes:
      - application/json
      description: update role information by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role information
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateRoleByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdateRoleByIDRespond'
      security:
      - BearerAuth: []
      summary: update role
      tags:
      - role
  /api/v1/role/condition:
    post:
      consumes:
      - application/json
      description: get role by condition
      parameters:
      - description: query condition
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Conditions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetRoleByConditionRespond'
      security:
      - BearerAuth: []
      summary: get role by condition
      tags:
      - role
  /api/v1/role/delete/ids:
    post:
      consumes:
      - application/json
      description: delete roles by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.DeleteRolesByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteRolesByIDsRespond'
      security:
      - BearerAuth: []
      summary: delete roles
      tags:
      - role
  /api/v1/role/list:
    get:
      consumes:
      - application/json
      description: list of roles by last id and limit
      parameters:
      - default: 0
        description: last id, default is MaxInt32
        in: query
        name: lastID
        required: true
        type: integer
      - default: 10
        description: size in each page
        in: query
        name: limit
        type: integer
      - default: -id
        description: 'sort by column name of table, and the '
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListRolesRespond'
      security:
      - BearerAuth: []
      summary: list of roles by last id and limit
      tags:
      - role
    post:
      consumes:
      - application/json
      description: list of roles by paging and conditions
      parameters:
      - description: query parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Params'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListRolesRespond'
      security:
      - BearerAuth: []
      summary: list of roles by query parameters
      tags:
      - role
  /api/v1/role/list/ids:
    post:
      consumes:
      - application/json
      description: list of roles by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.ListRolesByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListRolesByIDsRespond'
      security:
      - BearerAuth: []
      summary: list of roles by batch id
      tags:
      - role
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type Bearer your-jwt-token to Value
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"